[SUPPORT-1805] Change offset to long time to allow WARC files bigger than 4 GB. Kept length as an int as it's unlikely we'll stream 4GB of data through a browser (cached copies).
git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@42855 3cdb6afb-a121-0410-a22c-8da0070c90a1

+++ publicui-web/src/main/java/com/funnelback/publicui/search/service/DataRepository.java
@@ -29,7 +29,7 @@ public interface DataRepository {
      * @return Document The document
      */
     public RecordAndMetadata<? extends Record<?>> getDocument(Collection collection, StoreView view, String url,
-        File relativePath, int offset, int length);
+        File relativePath, long offset, int length);
     
     /**
      * Gets a cached document.
